Converted repo into a Houdini package
Lightly refactored this repo to enable  it for use as a Houdini package.

Downloading and installing pydicom is no longer required as the required libraries are included within the repo, which gets appended to the Houdini 'PYTHONPATH''

Updated the file, 'readme.md', with installation and usage instructions for the previously stated alterations.